Cambios
creado fichero para subir mediciones
modificado el get sondas para que devuelva el nuevo campo notificacion y creada la funcionalidad de comprobar notificaciones, creada la funcionalidad de cuando cierres sesion que te vuelva a pedir de iniciar sesion como le gusta a JL

+++ src/js/login.js
@@ -1,7 +1,7 @@
 var path;
-if (!location.pathname.includes("/app/")){
+if (!location.pathname.includes("/app/")) {
     path = "./"
-}else{
+} else {
     path = "../"
 }
 
@@ -10,7 +10,7 @@ document.querySelector("form").addEventListener("submit", function (event) {
     event.preventDefault()
     let dataLogin = new FormData(event.target);
     dataLogin.set('recordar', document.getElementById("recuerdame").checked.toString())
-    let url=path+"api/v1.0/" + "sesion"
+    let url = path + "api/v1.0/" + "sesion"
     fetch(url, {
         method: "POST",
         body: dataLogin,
@@ -19,32 +19,79 @@ document.querySelector("form").addEventListener("submit", function (event) {
             return respuesta.json()
         }
     }).then(function (datos) {
-         if (datos['rol'] === "usuario")
-             location = path+"app/miscampos.php"
-         if (datos['rol'] === "admin")
-             location = "."
+        checkNotificaciones(datos['id'])
+        setTimeout(function () {
+            if (datos['rol'] === "usuario")
+                location = path + "app/miscampos.php"
+            if (datos['rol'] === "admin")
+                location = "."
+        },200)
 
 
     }).catch(function (error) {
-         // document.getElementById("picuser").setAttribute("onclick", "loginBlur(true)")
-         console.log(error)
-         document.getElementById("output").textContent = "Las credenciales son incorrectas"
+        // document.getElementById("picuser").setAttribute("onclick", "loginBlur(true)")
+        console.log(error)
+        document.getElementById("output").textContent = "Las credenciales son incorrectas"
 
     })
 })
 
 
 function logout() {
-    fetch(path+'api/v1.0/sesion', {
+    fetch(path + 'api/v1.0/sesion', {
         method: 'DELETE',
-    }).then(function(respuesta) {
+    }).then(function (respuesta) {
         if (respuesta.ok) {
             var opcion = confirm("¿Estás seguro de que quieres cerrar sesión?")
-            if (!opcion){
+            if (!opcion) {
                 return
             }
-            sessionStorage.clear();
-            location.href = path+'index.php';
+            if (opcion){
+                sessionStorage.clear();
+                location.href = path + 'index.php';
+                abrirIniciosesion(true)
+            }
         }
+    })
+}
+
+function abrirIniciosesion(acitvar) {
+    sessionStorage.setItem("activar", acitvar)
+
+
+
+}
+
+function checkNotificaciones(idUser) {
+    fetch(path + "api/v1.0/parcelas/" + idUser, {
+        method: 'GET'
+    }).then(function (respuesta) {
+        if (respuesta.ok) {
+            return respuesta.json()
+        }
+    }).then(function (data) {
+        setTimeout(function () {
+            console.log(data)
+        },3000)
+
+        for (const campo of data) {
+            fetch(path + "api/v1.0/sondas/" + campo['id'], {
+                method: 'GET'
+            }).then(function (respuesta) {
+                if (respuesta.ok) {
+                    return respuesta.json()
+                }
+            }).then(function (data) {
+                console.log("hola")
+                for (const sonda of data) {
+                    if (sonda['notificacion'] != '') {
+                        document.getElementById("contenedor-header").setAttribute('style', "background:red")
+                        break;
+                    }
+                }
+            })
+
+        }
+        console.log("hola")
     })
 }
